+++ src/app/(dashboard)/profile/[id]/page.tsx
@@ -1,7 +1,6 @@
 import NotFound from "@/src/components/Dashboard/NotFound/NotFound"
 import ProfileActivities from "@/src/components/Dashboard/profile/profile-activities"
 import ProfileBio from "@/src/components/Dashboard/profile/profile-bio"
-// import ProfileCalendar from "@/src/components/Dashboard/profile/profile-calendar"
 import ProfileRewards from "@/src/components/Dashboard/profile/profile-rewards"
 import { ExtendedRecommendations } from "@/src/components/Dashboard/profile/types/profile-types"
 import ProfileFollowActions from "@/src/components/Dashboard/profile/user/ProfileFollowActions"
